refactor(runtime-api): extract workspace status helpers
